Moved to clang++
It quite some time that I play with clang, and finally I've decided to use it
as official compiler.
Error reporting is so much clear and compile time is much faster.
Up to now flags are exactely the same as g++.

Added a small define to avoid the "unused variable error" for those cases
that I want unused variable around.