Game/Combat engine first steps
This commit is just a first step in the creation of the
RPG game system and combat system.

It is difficult to say how it will endup in the end, but
I'm trying to pin down design ideas by writing mockup classes
and hierarchy:inheritance.

The main ideas are:

Entity: anythig that exist as an object/terrain/creature/trap
Creature -> Entity: A specific type of entity (has physical/mental)
        attributes

Effect: anything that change/influence an entity:
        spell, disease, environmental effects (wind)
        * At this point not sure if Wounds/Fatigue should be an effect.

Combat_Engine: a class to which you pass two cratures and will change
               them based on combat result.
               * This is just a place holder and WILL change