Implemeting dice roller and stats
Continuing to implement the Dice class and associated Creature class.
Added some skeleton roll methods that will grow in something with a bit
more functionality.

The idea is to have a Dice roller that can be called everywhere exporting
more or less generic services.

Started to populate the stats/skills map: ideally some lua file should
be used to create the creatures and populate these structures