Extended game table initialization
I've rearranged some of the tables/resources needed to initialized the game
engine.

I've added a first version for the combat matrices, defining all the outcome
and creating a multi dimensiona vector (interesting experiment).

It still remain to define the hundreds of resutls in the dozens of tables:
it will take time but now the structure is in place.